Move SoraniNormalizationFilterFactory to the common analysis plugin (#33892)
Follow up to #25715
